chore: refactor post/read
