+++ modules/cached-resources/index.js
@@ -1,0 +1,16 @@
+import { Form } from 'formik'
+import { residentQuery } from './read';
+import {
+  postIdentificationForm,
+  postSupplementaryForm,
+  postHousehold,
+  postHouseholdWithRelation
+} from './post';
+
+export {
+  residentQuery,
+  postIdentificationForm,
+  postSupplementaryForm,
+  postHousehold,
+  postHouseholdWithRelation
+}
